Moved status test to the generic AnyOperatorTestCase
